Fixing issue with properties not being the format of a Map[String, String]
